Make `add_mod()` commutative and document the range of accepted values
The previous version of `add_mod(x, y, m)` implicitly assumed that
`x < m`, `y <= m`, and `m > 0`. While these assumptions are correct,
and the code respects them, they may generate confusion in the future
(GH-1). In particular, one might expect `add_mod(x, y, m)` to be the
same as `add_mod(y, x, m)`, but this may not be the case if `y = m`.

This commit:
- changes the assumptions to `x <= m`, `y <= m`, and `m > 0`
- documents these assumptions explicitly, and enforces them through
  debug assertions
- adds tests for `add_mod()`