Fix `strict-provenance` error in `swap()`
The call to `get_maybe_uninit()` was wrong, because a cast from
`&const _` to `&mut _` is UB. But also a using `get_maybe_uninit_mut()`
would be wrong, because the seconds invocation of the method invalidates
the exclusive borrow of the first call.

This changes inlines the functionality of `get_maybe_uninit_mut()`, to
fix the problem. It can be tested with:

```sh
MIRIFLAGS="-Zmiri-strict-provenance" cargo +nightly miri test
```