Make the `unstable` feature working again
Indexing slices with ranges (`slice[start..end]`) was supported as a
`const` operation at a certain point with some older versions of Rust,
but this hasn't been the case for a long time now. This commit changes
all the unstable `const fn` to `fn` because of that.

This commit also removes some `#[feature]` directives that are now
unnecessary.

Now the crate can be successfully built using `--features unstable` on
the latest Rust nightly compiler.