Invert the order of generics in the `CircularBuffer` struct
Changed from:

    CircularBuffer<const N: usize, T>

to:

    CircularBuffer<T, const N: usize>

so that now `CircularBuffer<T, N>` is more consistent with the standard
array type `[T; N]`.

The only reason why `CircularBuffer` had the `N` as the first parameter
is that it was created back when const generics were unstable, and at
that time const generics could only appear before generic types. The
previous order of parameters was never an intentional decision.